requirements.txt
PyYAML~=5.3.1
torch~=1.6.0
torchvision~=0.7.0
pillow~=8.0.1
rux
torchsummary~=1.5.1
runx~=0.0.10